+++ frontend/src/components/Header.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class Header extends Component {
+    render() {
+        return(
+            <header>
+                JARP REACT COMMUNITY VIDEO PLAYER
+            </header>
+        );
+    }
+}
+
+export default Header;